requirements.txt
langchain
python-dotenv
langchain_community
ipykernel
pypdf
bs4
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
fastapi
uvicorn
langserve
arxiv
wikipedia
ddgs